Split things into modules
It's going to make it nicer when we'll have a lot of functions
in the task module.
